split reserved gathers by which lair does it
Allow reserving gathers for blue and orange lair separately via
`--reserve-gathers-blue` and `--reserve-gathers-orange`. This is useful
when planning for each to reserve gathers through a different finder and
maybe also with a different range limit.

Also remove the need to specify lair actions in cmdline args.
`lair_blue` and `lair_orange` are always included in the permutated set.

Closes #11
Signed-off-by: Tal Shorer <tal.shorer@gmail.com>