trace action tree to its progenitor action in cat cafe output
When an action triggers another action or more in a chain, output it in
cat cafe as if the original action make all the modifications. This is
important in a case that invaders move to a portal or get added by
Military Response, as what we actually care about is not the triggered
action but the action that originally moved or destroyed them.

Fixes #42
Signed-off-by: Tal Shorer <tal.shorer@gmail.com>