lrlr: leave invaders in range-1 lands with dahan during slurp if there are enough ravages left to wipe them out
When leaving invaders in range-1 lands with dahan, those invaders will
MR again in-place, preventing us from clearing the land. Decide which
invaders are OK to leave in range-1 lansd with dahan based on their
health. During every MR, a new invader with one less health is added. If
an invader's health is less than the expected number of ravages left,
which are expected to completely wipe everything at range-1, that
invader can be safely left at range-1 to be killed as many times as
needed.

Fixes #91
Signed-off-by: Tal Shorer <tal.shorer@gmail.com>