adjacency: attempt to rotate again for weird corner configurations
Sometimes, there can be four lands touching at a corner (call them lands
W-X-Y-Z), such that W and Z are on edges that aren't connected. They
should still be adjacent, but that can only be found if we attempt
rotation twice in the same direction, since antirotation yields an
unconnected edge.

An example of where this happens in the 144p map is a spoke's U CLOCK11
corner and its adjacent rim's Q CLOCK5 corner.

Signed-off-by: Tal Shorer <tal.shorer@gmail.com>